refactor: Adding required containers and moving unwanted logic from pages
