+++ client/src/containers/LoginContainer/index.ts
@@ -1,0 +1,1 @@
+export { LoginContainer } from "./Login.container";
