+++ client/src/containers/NewNoteContainer/index.ts
@@ -1,0 +1,1 @@
+export { NewNoteContainer } from "./NewNote.container";
