+++ client/src/containers/NoteListContainer/index.ts
@@ -1,0 +1,1 @@
+export { NoteListContainer } from "./NoteList.container";
