+++ client/src/containers/SignupContainer/index.ts
@@ -1,0 +1,1 @@
+export { SignupContainer } from "./Signup.container";
