+++ client/src/containers/index.ts
@@ -1,2 +1,6 @@
 export { ForgotPasswordContainer } from "./ForgotPasswordContainer";
+export { LoginContainer } from "./LoginContainer";
+export { NewNoteContainer } from "./NewNoteContainer";
+export { NoteListContainer } from "./NoteListContainer";
 export { ResetPasswordContainer } from "./ResetPasswordContainer";
+export { SignupContainer } from "./SignupContainer";
