--- client/src/pages/login/LoginPage.types.ts
@@ -1,5 +1,0 @@
-export interface LoginFormFields {
-  email: string;
-  password: string;
-  remember: boolean;
-}
