--- client/src/pages/new-note/NewNote.types.ts
@@ -1,6 +1,0 @@
-import { Note } from "@/types";
-
-export type NewNoteFormData = Omit<
-  Note,
-  "id" | "in_trash" | "user_id" | "created_at"
->;
